Extract FileProvider into Frost.Effects.FileProvider
